backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
sentence-transformers
transformers
torch
faiss-cpu
scikit-learn
